requirements.txt
# Core dependencies
groq==0.22.0
gTTS==2.5.4
python-dotenv==1.0.0
tqdm==4.67.1

# GUI dependencies (optional)
PyQt5==5.15.9

# Development and testing dependencies
pytest==7.4.3
pytest-mock==3.12.0
mypy==1.7.1
flake8==6.1.0

# Additional utilities
requests==2.31.0
pathlib2==2.3.7
